refactor: modify service for this endpoint
