+++ src/main/java/tech/gruppone/stalker/server/services/ConnectionsService.java
@@ -1,0 +1,44 @@
+package tech.gruppone.stalker.server.services;
+
+import java.sql.Timestamp;
+import org.springframework.stereotype.Service;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.experimental.FieldDefaults;
+import reactor.core.publisher.Flux;
+import tech.gruppone.stalker.server.model.api.UserDataDto;
+import tech.gruppone.stalker.server.model.api.UserDto;
+import tech.gruppone.stalker.server.repositories.UserDataRepository;
+import tech.gruppone.stalker.server.repositories.UserRepository;
+
+@AllArgsConstructor
+@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
+@Service
+
+public class ConnectionsService {
+  UserRepository userRepository;
+  UserDataRepository userDataRepository;
+
+  public Flux<UserDto> findConnectedUsersByOrganizationId(Long organizationId) {
+
+    return userRepository
+        .findAllUsers(organizationId)
+        .zipWith(userDataRepository.findAllUserData(organizationId))
+        .map(
+            result -> {
+              var t1 = result.getT1();
+              var t2 = result.getT2();
+              return UserDto.builder()
+                  .id(t1.getId())
+                  .data(
+                      UserDataDto.builder()
+                          .email(t1.getEmail())
+                          .firstName(t2.getFirstName())
+                          .lastName(t2.getLastName())
+                          .birthDate(t2.getBirthDate())
+                          .creationDateTime(Timestamp.valueOf(t2.getCreatedDate()))
+                          .build())
+                  .build();
+            });
+  }
+}
