+++ src/main/java/tech/gruppone/stalker/server/services/OrganizationService.java
@@ -13,12 +13,8 @@ import tech.gruppone.stalker.server.model.api.OrganizationDataDto.OrganizationDa
 import tech.gruppone.stalker.server.model.api.OrganizationDto;
 import tech.gruppone.stalker.server.model.api.PlaceDataDto;
 import tech.gruppone.stalker.server.model.api.PlaceDto;
-import tech.gruppone.stalker.server.model.api.UserDataDto;
-import tech.gruppone.stalker.server.model.api.UserDto;
 import tech.gruppone.stalker.server.model.db.OrganizationDao;
 import tech.gruppone.stalker.server.repositories.OrganizationRepository;
-import tech.gruppone.stalker.server.repositories.UserDataRepository;
-import tech.gruppone.stalker.server.repositories.UserRepository;
 
 @AllArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
@@ -27,8 +23,6 @@ import tech.gruppone.stalker.server.repositories.UserRepository;
 public class OrganizationService {
   OrganizationRepository organizationRepository;
   PlaceService placeService;
-  UserRepository userRepository;
-  UserDataRepository userDataRepository;
 
   public Flux<OrganizationDto> findAll() {
     // I have no idea why this works
@@ -74,28 +68,5 @@ public class OrganizationService {
         .doOnNext(orgDao -> log.info("created organization {}", orgDao))
         .map(OrganizationDao::getId)
         .doOnNext(organizationId -> placeService.saveAll(placeDataDtos, organizationId));
-  }
-
-  public Flux<UserDto> findConnectedUsersByOrganizationId(Long organizationId) {
-
-    return userRepository
-        .findAllUsers(organizationId)
-        .zipWith(userDataRepository.findAllUserData(organizationId))
-        .map(
-            result -> {
-              var t1 = result.getT1();
-              var t2 = result.getT2();
-              return UserDto.builder()
-                  .id(t1.getId())
-                  .data(
-                      UserDataDto.builder()
-                          .email(t1.getEmail())
-                          .firstName(t2.getFirstName())
-                          .lastName(t2.getLastName())
-                          .birthDate(t2.getBirthDate())
-                          .creationDateTime(Timestamp.valueOf(t2.getCreatedDate()))
-                          .build())
-                  .build();
-            });
   }
 }
